messed around and got a triple double
