+++ src/components/myAlbumAdd.js
@@ -8,9 +8,12 @@ import { requestFriend } from '../actions/friendActions';
 import { fetchUserInfo } from '../actions/userActions';
 import {Button} from 'react-native-paper';
 import { Icon } from 'react-native-elements';
+import axios from 'axios';
 
+const API = 'https://rekall-server.herokuapp.com';
 
 class MyAlbumRoute extends Component{
+
     constructor(props){
         super(props);
         this.state = {
@@ -35,6 +38,38 @@ class MyAlbumRoute extends Component{
         this.getCurrentAlbums();
     }
 
+    addToGallery = async () => {
+        var url = `${API}/album/addMediaToAlbum`
+        let promises = [];
+        
+        axios.put(`${API}/album/addMediaToLibrary`,
+            { 
+                "_id": this.props.user.uid,
+                "mediaURL": 'https://www.youtube.com/watch?v=' + this.props.video,
+                "mediaType": 'YouTube'
+            }).then((res) => {
+                let mediaid = res.data._id;
+
+                for(let i = 0; i < this.state.updateAlbums.length; i++) {
+                    promises.push(axios.put(url,
+                        { 
+                            "album": {
+                                "_id": this.state.updateAlbums[i].toString(),
+                            },
+                            "media": {
+                                "_id": mediaid,
+                            },
+                        },
+                    ));
+                }
+                Promise.all(promises).then(() => {
+                    this.loadData();
+                    this.props.navigation.goBack();
+                })
+            }).catch((e) => {
+                console.log(`Error putting media: ` + JSON.stringify(e));
+            });
+    }
 
     getCurrentAlbums(){
         var albumList = this.state.albumList;
@@ -61,7 +96,6 @@ class MyAlbumRoute extends Component{
                 </View>
             )
         }
-
 
         return(
             <LinearGradient
@@ -92,7 +126,7 @@ class MyAlbumRoute extends Component{
                             />            
                         </View>
                         <View style={styles.addButtonBox}>
-                            <Button mode='contained'  color="#F2F1F1" labelStyle={styles.buttonText} onPress={() => this.addToGallery('User')} >
+                            <Button mode='contained'  color="#F2F1F1" labelStyle={styles.buttonText} onPress={() => this.addToGallery()} >
                                 Add to Gallery
                             </Button>
                         </View>
