+++ src/components/sharedAlbumAdd.js
@@ -7,8 +7,9 @@ import DropDownPicker from 'react-native-dropdown-picker';
 import { requestFriend } from '../actions/friendActions';
 import { fetchUserInfo } from '../actions/userActions';
 import {Button} from 'react-native-paper';
+import axios from 'axios';
 
-
+const API = 'https://rekall-server.herokuapp.com';
 
 class SharedAlbumRoute extends Component{
     constructor(props){
@@ -32,6 +33,40 @@ class SharedAlbumRoute extends Component{
             });
         });
         this.getCurrentAlbums();
+    }
+
+    addToGallery = async () => {
+        console.log(this.state.updateAlbums);
+        var url = `${API}/album/addMediaToShared`
+        let promises = [];
+        
+        axios.put(`${API}/album/addMediaToLibrary`,
+            { 
+                "_id": this.props.user.uid,
+                "mediaURL": 'https://www.youtube.com/watch?v=' + this.props.video,
+                "mediaType": 'YouTube'
+            }).then((res) => {
+                let mediaid = res.data._id;
+
+                for(let i = 0; i < this.state.updateAlbums.length; i++) {
+                    promises.push(axios.put(url,
+                        { 
+                            "album": {
+                                "_id": this.state.updateAlbums[i].toString(),
+                            },
+                            "media": {
+                                "_id": mediaid,
+                            },
+                        },
+                    ));
+                }
+                Promise.all(promises).then(() => {
+                    this.loadData();
+                    this.props.navigation.navigate('Drawer', { screen: 'EXPLORE' });
+                })
+            }).catch((e) => {
+                console.log(`Error putting media: ` + JSON.stringify(e));
+            });
     }
 
     getCurrentAlbums(){
@@ -58,7 +93,7 @@ class SharedAlbumRoute extends Component{
                 </View>
             )
         }
-
+        
         return(
             <LinearGradient
             colors={['#FFFFFF', '#D9D9D9']}
