+++ src/components/videoDetailTabs.js
@@ -18,18 +18,24 @@ const renderTabBar = props => (
 
 const initialLayout = { width: Dimensions.get('window').width };
  
-const VideoDetailTabs =(updateAlbums, myalbumNames, sharedalbumNames) => {
+const VideoDetailTabs =(props) => {
   const [index, setIndex] = React.useState(0);
   const [routes] = React.useState([
-    { key: 'first', title: 'My Albums', updateAlbums: updateAlbums, albums: myalbumNames},
-    { key: 'second', title: 'Shared Albums', updateAlbums: updateAlbums, albums: sharedalbumNames},
+    { key: 'first', title: 'My Albums'},//, updateAlbums: updateAlbums, albums: myalbumNames, video: videoid},
+    { key: 'second', title: 'Shared Albums'}//, updateAlbums: updateAlbums, albums: sharedalbumNames, video: videoid},
   ]);
- 
-  const renderScene = SceneMap({
-    first: MyAlbumRoute,
-    second: SharedAlbumRoute,
-  });
- 
+
+  const renderScene = ({ route }) => {
+    switch (route.key) {
+      case 'first':
+        return <MyAlbumRoute video={props.videoid} />;
+      case 'second':
+        return <SharedAlbumRoute video={props.videoid} />;
+      default:
+        return null;
+    }
+  };
+
   return (
     <TabView
       navigationState={{ index, routes }}
